chore(source): free space whenever possible
Whenever a notification API response is parsed , all notification
last read and id are concatnated to form a unique string , this
string is stored in a QStringList in the heap. In the older versions
the program did not care to free unused data. In the future this
may lead to memory overhaul where unused data fills up all the heap.
Also the memory usage will increase propotional to the usage time
which is not desirable. This commit changes. Whenever the github
api returns an empty notifications response , We free the QStringList
and so the memory usage will not go above a certain threshold set
by github.(I guess)